Correct MAX limit constant.

+++ xmltooling/XMLToolingConfig.cpp
@@ -368,7 +368,7 @@ bool XMLToolingInternalConfig::init()
 
     Lock initLock(m_lock);
 
-    if (m_initCount == LONG_MAX) {
+    if (m_initCount == INT_MAX) {
         log.crit("library initialized too many times");
         return false;
     }
